+ working materialRegistry (on load) - rename Unifier to MaterialManager and move static method to there - make API package clean from implementation class as possible
