+++ src/main/java/mods/bio/gttweaker/api/oredict/IMaterialData.java
@@ -1,0 +1,22 @@
+package mods.bio.gttweaker.api.oredict;
+
+import stanhebben.zenscript.annotations.ZenClass;
+import stanhebben.zenscript.annotations.ZenGetter;
+
+import java.util.List;
+
+@ZenClass("mods.gregtech.oredict.IMaterialData")
+public interface IMaterialData {
+
+	@ZenGetter
+	IMaterialStack material();
+
+	@ZenGetter
+	IPrefix prefix();
+
+	@ZenGetter
+	List<IMaterialStack> byProducts();
+
+	@ZenGetter
+	List<IMaterialStack> materials();
+}
