+++ src/main/java/mods/bio/gttweaker/gregtech/oredict/CTMaterial.java
@@ -1,0 +1,57 @@
+package mods.bio.gttweaker.gregtech.oredict;
+
+import gregapi.oredict.OreDictMaterial;
+import gregapi.oredict.OreDictMaterialStack;
+import minetweaker.MineTweakerAPI;
+import mods.bio.gttweaker.api.oredict.IMaterial;
+import mods.bio.gttweaker.api.oredict.IMaterialFactory;
+import mods.bio.gttweaker.api.oredict.IMaterialStack;
+
+import java.util.Objects;
+
+public class CTMaterial implements IMaterial {
+	private final OreDictMaterial material_internal;
+
+	@Override
+	public OreDictMaterial getMaterial(){
+		return material_internal;
+	}
+
+	@Override
+	public IMaterialFactory edit() {
+		return new CTMaterialFactory(this);
+	}
+
+	@Override
+	public IMaterialStack multiply(long amount) {
+		return new CTMaterialStack(new OreDictMaterialStack(getMaterial(),amount));
+	}
+
+	public CTMaterial(OreDictMaterial aMaterial){
+		if (aMaterial == null){
+			MineTweakerAPI.logError("Null Material was provided unable to create a new <material:>");
+		}
+		material_internal = aMaterial;
+	}
+
+	/**
+	 * @return formatted string as it is formatted in bracketHandler
+	 */
+	@Override
+	public String toString() {
+		return String.format("<material:%s>",material_internal.mNameInternal);
+	}
+
+	@Override
+	public boolean equals(Object o) {
+		if (this == o) return true;
+		if (o == null || getClass() != o.getClass()) return false;
+		CTMaterial that = (CTMaterial) o;
+		return Objects.equals(material_internal, that.material_internal);
+	}
+
+	@Override
+	public int hashCode() {
+		return material_internal.hashCode();
+	}
+}
